+++ shelfsense/src/pages/ProductsPage.jsx
@@ -4,7 +4,7 @@ import {Button, Box, CircularProgress, Tooltip} from '@mui/material';
 import ProductsList from '../components/products/ProductsList';
 import { Sidebar } from '../components/sidebar/sidebar.jsx';
 import ProductCreateModal from '../components/products/ProductCreateModal.jsx';
-import {createApiRequest} from "../util/services/ComponentService.jsx";
+import {createApiRequest} from "../util/services/components/createApiRequest.js";
 import useComponentsStore from "../stores/useComponentsStore.js";
 import useApiUpdateStore from "../stores/useApiUpdateStore.js"; // Import the modal
 import calculateApiFetchTimeDif from '../util/calculateApiFetchTimeDif.js';
