+++ shelfsense/src/util/services/ComponentService.jsx
@@ -1,53 +1,2 @@
-import useSessionStore from "../../stores/useSessionStore.js";
-import useApiUpdateStore from "../../stores/useApiUpdateStore.js";
-import {destroyStoresAndLogout} from "../destroyStoresAndLogout.js";
 
-const BASE_URL = `${import.meta.env.VITE_API_URL}/components`;
-
-export const createApiRequest = async () => {
-
-    const user = useSessionStore.getState().user;
-
-    const sendApiRequest = {
-        userId: user.id,
-        apiKey: import.meta.env.VITE_API_KEY,
-    };
-
-    console.log(sendApiRequest);
-
-    try{
-        const response = await fetch(`${BASE_URL}/mouser`, {
-            method: "POST",
-            headers: {"Content-Type": "application/json",},
-            credentials: "include",
-            body: JSON.stringify(sendApiRequest),
-        });
-
-        if (response.status === 401) {
-            await destroyStoresAndLogout();
-            return null;
-        }
-
-        if (!response.ok) {
-            console.error('Failed to create component:', response.status);
-            return null;
-        }
-
-        const getLocalDateTime = () => {
-            const now = new Date();
-            return now.toISOString();
-        };
-
-        useApiUpdateStore.getState().setApiUpdate({ lastUpdated: getLocalDateTime() });
-
-        console.log(getLocalDateTime());
-
-        console.log("Last Updated: ", useApiUpdateStore.getState().apiUpdate);
-        console.log("Time now: ", getLocalDateTime());
-        return await response.json();
-    } catch (error) {
-        console.error('Error occurred while creating component:', error);
-        return null;
-    }
-}
