+++ shelfsense/src/util/services/components/createComponent.js
@@ -1,8 +1,8 @@
 import {destroyStoresAndLogout} from "../../destroyStoresAndLogout.js";
 
-const BASE_URL = `${import.meta.env.VITE_API_URL}/components`;
 
 export const createComponent = async (componentData) => {
+    const BASE_URL = `${import.meta.env.VITE_API_URL}/components`;
 
     try {
         const response = await fetch(BASE_URL, {
